requirements.txt
pyTelegramBotAPI==4.14.0     # or latest compatible version
python-dotenv==1.0.1         # for loading .env files
oauth2client==4.1.3          # required by gspread for auth
gspread==5.12.0              # Google Sheets API
gspread-formatting==0.5.0    # Google sheet formatting (coloring)